Let all spawned Surface obey to 'context.editable' setting.
